+++ src/preload.ts
@@ -50,7 +50,7 @@ import { fixHotReloadNext12 } from './utils/preload';
 import { isFolder, getDirFiles } from './api/fs';
 import { convertHEIC } from './api/heicConvert';
 import { setupLogging } from './utils/logging';
-import { setupRendererProcessStatsLogger } from './utils/memory';
+import { setupRendererProcessStatsLogger } from './utils/processStats';
 import { runFFmpegCmd } from './api/ffmpeg';
 
 fixHotReloadNext12();
